# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampusMatch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get campus feed
      parameters:
        - in: query
          name: category
          schema:
            type: ["string", "null"]
            enum: [general, study, sports, hackathon, events, null]
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                category:
                  type: string
                imageUrl:
                  type: ["string", "null"]
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Toggle like on a post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── EVENTS ────────────────────────────────────────────────────────────────────
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get campus events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event (admin)
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date-time
                venue:
                  type: string
                category:
                  type: string
                imageUrl:
                  type: ["string", "null"]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (admin)
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /events/{eventId}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RSVP recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── MARKETPLACE ───────────────────────────────────────────────────────────────
  /marketplace:
    get:
      operationId: getListings
      tags: [marketplace]
      summary: Get marketplace listings
      parameters:
        - in: query
          name: category
          schema:
            type: ["string", "null"]
            enum: [books, cycles, electronics, furniture, other, null]
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [marketplace]
      summary: Create a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /marketplace/{listingId}:
    patch:
      operationId: updateListing
      tags: [marketplace]
      summary: Update a listing
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                price:
                  type: integer
                category:
                  type: string
                imageUrl:
                  type: ["string", "null"]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [marketplace]
      summary: Delete a listing
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /marketplace/{listingId}/sold:
    put:
      operationId: toggleListingSold
      tags: [marketplace]
      summary: Toggle listing sold status
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"

  # ── UPLOADS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required: [success]

    # ── AUTH ──────────────────────────────────────────────────────────────────
...
    Post:
      type: object
      required: [id, userId, content, category, likeCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        likeCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [content, category]
      properties:
        content:
          type: string
          minLength: 1
        category:
          type: string
          enum: [general, study, sports, hackathon, events]
        imageUrl:
          type: ["string", "null"]

    # ── EVENTS ────────────────────────────────────────────────────────────────
    Event:
      type: object
      required: [id, title, description, date, venue, category, rsvpCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        venue:
          type: string
        category:
          type: string
          enum: [hackathon, sports, cultural, academic, social]
        imageUrl:
          type: ["string", "null"]
        rsvpCount:
          type: integer
        isRsvped:
          type: boolean
        createdAt:
          type: string
          format: date-time
    EventInput:
      type: object
      required: [title, description, date, venue, category]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        venue:
          type: string
        category:
          type: string
          enum: [hackathon, sports, cultural, academic, social]
        imageUrl:
          type: ["string", "null"]

    # ── MARKETPLACE ───────────────────────────────────────────────────────────
    Listing:
      type: object
      required: [id, userId, title, description, price, category, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        sellerName:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: string
          enum: [books, cycles, electronics, furniture, other]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, sold]
        createdAt:
          type: string
          format: date-time
    ListingInput:
      type: object
      required: [title, description, price, category]
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: integer
        category:
          type: string
          enum: [books, cycles, electronics, furniture, other]
        imageUrl:
          type: ["string", "null"]

    # ── ANNOUNCEMENTS ─────────────────────────────────────────────────────────
...